arregla bug overflow

+++ pap-tp1/ej1/ej1.cpp
@@ -4,10 +4,10 @@
 
 using namespace std;
 
-typedef vector<int>::iterator it;
+typedef vector<int64_t>::iterator it;
 
-int suma(it begin, it end, int64_t mask) {
-    int s = 0;
+int64_t suma(it begin, it end, int64_t mask) {
+    int64_t s = 0;
     for (auto d = begin; d != end; ++d) {
         if (mask & (1 << (d - begin))) {
             s += *d;
@@ -16,10 +16,10 @@ int suma(it begin, it end, int64_t mask) {
     return s;
 }
 
-vector<int> soluciones(it begin, it end, int p) {
-    vector<int> sol;
+vector<int64_t> soluciones(it begin, it end, int64_t p) {
+    vector<int64_t> sol;
     for (int64_t i = 0; i < (1 << (end - begin)); i++) {
-        int s = suma(begin, end, i);
+        int64_t s = suma(begin, end, i);
         if (s <= p) {
             sol.push_back(s);
         }
@@ -28,12 +28,12 @@ vector<int> soluciones(it begin, it end, int p) {
 }
 
 int main() {
-    int p, n;
+    int64_t p, n;
     cin >> p >> n;
 
-    vector<int> ds(n);
+    vector<int64_t> ds(n);
 
-    for (int& d : ds) cin >> d;
+    for (int64_t& d : ds) cin >> d;
 
     auto sol1 = soluciones(ds.begin(), ds.begin() + n/2, p);
     auto sol2 = soluciones(ds.begin() + n/2, ds.end(), p);
@@ -44,10 +44,10 @@ int main() {
     auto it1 = sol1.begin();
     auto it2 = sol2.rbegin();
 
-    int max_sum = 0;
+    int64_t max_sum = 0;
 
     while (it1 != sol1.end() && it2 != sol2.rend()) {
-        int sum = (*it1) + (*it2);
+        int64_t sum = (*it1) + (*it2);
         if (sum <= p) { // = p shortcut
             if (sum > max_sum) {
                 max_sum = sum;
